Updating package.json to include an option for coverage, had to re-write some stuff to make sure that views were not getting coverage statements inserted (breaking things).
